Reutilizar trend actions mgs en trend-detail (delete)

+++ src/app/modules/trends/trend-detail/trend-detail.component.ts
@@ -6,11 +6,16 @@ import { SidenavEndService } from 'src/app/modules/sidenav-end/services/sidenav-
 import { selectSelectedTrend } from '../store/selectors';
 import { Trend } from '../models/trend.model';
 import { deleteOneTrend } from '../store/actions/trend-crud.actions';
-import { selectactionRequireTrendState } from '../store/reducers';
+import { selectMessageState, selectactionRequireTrendState } from '../store/reducers';
 import { TrendActionEnum } from '../enums/trend-actions.enum';
-import { actionRequireTrendEditState } from '../store/actions/trends-list-page.actions';
+import { actionRequireTrendEditState, updateMessageTrendState } from '../store/actions/trends-list-page.actions';
 import { Router, RouterLink } from '@angular/router';
 import { NgIf, NgFor, AsyncPipe } from '@angular/common';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../dialogs/components/confirm-dialog/confirm-dialog.component';
+import { TrendMsgAction } from '../models/trend-msg-action.model';
+import { TrendService } from '../services/trend.service';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 @Component({
     selector: 'app-trend-detail',
@@ -22,11 +27,14 @@ import { NgIf, NgFor, AsyncPipe } from '@angular/common';
         NgIf,
         NgFor,
         AsyncPipe,
+        MatDialogModule,
+        MatSnackBarModule,
     ],
 })
 export class TrendDetailComponent implements OnInit, OnDestroy{
   protected trend$ = this.store.select(selectSelectedTrend);
   actionRequire$ = this.store.select(selectactionRequireTrendState);
+  messageState$ = this.store.select(selectMessageState);
 
   subscriptions: Subscription[] = [];
 
@@ -34,16 +42,34 @@ export class TrendDetailComponent implements OnInit, OnDestroy{
     private store: Store,
     private _sidenavEndService: SidenavEndService,
     private router: Router,
+    private dialog: MatDialog,
+    private _trendService: TrendService,
     ) {}
   ngOnDestroy(): void {
     this.subscriptions?.map((subs: Subscription) => subs.unsubscribe());
   }
-  ngOnInit(): void {
-    this.initSubscriptions();
-  }
+  ngOnInit(): void {}
 
   deleteTrend(trend: Trend) {
-    this.store.dispatch(deleteOneTrend({trendId: trend.id}));
+    const dialogDelete = this.dialog.open(ConfirmDialogComponent, {
+      width: '500px',
+      data: {
+        title: '¿Seguro que quieres eliminar la noticia?',
+        accept_label: 'Eliminar',
+        dismiss_label: 'Cancelar'
+      }
+    })
+    dialogDelete.componentInstance.onResult.subscribe((confirm) => {
+      if (confirm) {
+        this.initSubscriptions();
+        this.store.dispatch(deleteOneTrend({trendId: trend.id}));
+      }
+    }
+    );
+    dialogDelete.afterClosed().subscribe((_) => {
+      this.subscriptions?.map((subs: Subscription) => subs.unsubscribe());
+      dialogDelete.componentInstance.onResult.unsubscribe()
+    });
   }
   editTrend(trend: Trend) {
     this._sidenavEndService.overlayActionSource.next({
@@ -61,6 +87,12 @@ export class TrendDetailComponent implements OnInit, OnDestroy{
           this.store.dispatch(actionRequireTrendEditState({action: null}));
           // AL eliminar un trend, navego al home
           this.router.navigate(['/']);
+        }
+      }),
+      this.messageState$.subscribe((message: TrendMsgAction) => {
+        if (message) {
+        this.store.dispatch(updateMessageTrendState({ msg: null }));
+        this._trendService.manageTrendAction(message);
         }
       }),
     );
